feat: add ExistChildren method to Metadata repository
- Add ExistChildren method to check if a parent Metadata code has any children
- Use this check to prevent the deletion of codes that have children

+++ PrimeAPI/Application/Service/ProductMetadataService.cs
@@ -75,6 +75,12 @@ namespace PrimeAPI.Application.Service
         public async Task DeleteAsync(int id)
         {
             if(id <= 0) throw new ArgumentException("Id must be greater than zero.", nameof(id));
+
+            if (await _repository.ExistChildren(id))
+            {
+                throw new InvalidOperationException($"Cannot Delete Entity with ID: {id} as the child exists");
+            }
+
             await _repository.DeleteAsync(id);
         }
     }
